requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.3.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.2
click==8.1.8
coloredlogs==15.0.1
cryptography==43.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
fireworks-ai==0.15.11
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.25.1
kubernetes==31.0.0
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-community==0.3.14
langchain-core==0.3.29
langchain-fireworks==0.2.6
langchain-openai==0.3.0
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.5
langchainhub==0.1.21
langgraph==0.2.62
langgraph-api==0.0.16
langgraph-checkpoint==2.0.9
langgraph-cli==0.1.65
langgraph-sdk==0.1.51
langsmith==0.2.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.20.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.58.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.13
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
posthog==3.7.5
propcache==0.2.1
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
-e git+ssh://git@github.com/mterada1228/horses-reputation-llm.git@d98b6c53e9b7f39353e4052eaec3cbcf63eb9f5e#egg=react_agent&subdirectory=llm-api
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
selenium==4.27.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sse-starlette==2.1.3
starlette==0.41.3
streamlit==1.41.1
structlog==24.4.0
sympy==1.13.3
tavily-python==0.5.0
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.1
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
